backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pandas>=2.0.0
yfinance>=0.2.0
cachetools>=5.3.0
httpx>=0.26.0
lxml>=4.9.0
python-dateutil>=2.9.0
python-dotenv>=1.0.0
